[merge] fix dir path for merge
