--- src/main/java/EveryEgile/jaeyeon/BoardApplication.java
@@ -1,18 +1,0 @@
-package EveryEgile.jaeyeon;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
-
-@SpringBootApplication
-@EnableAutoConfiguration
-@EnableJpaAuditing
-	//	(exclude = {DataSourceAutoConfiguration.class})
-public class BoardApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(BoardApplication.class, args);
-	}
-
-}
