+++ src/main/java/org/everyagile/everyagile/everyagile/dto/ProjectDto.java
@@ -1,8 +1,8 @@
-package EveryEgile.jaeyeon.dto;
+package org.everyagile.everyagile.everyagile.dto;
 
 
-import EveryEgile.jaeyeon.entity.Project;
-import EveryEgile.jaeyeon.entity.Sprint;
+import org.everyagile.everyagile.everyagile.entity.Project;
+import org.everyagile.everyagile.everyagile.entity.Sprint;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
