+++ src/main/java/org/everyagile/everyagile/everyagile/repository/SprintRepository.java
@@ -1,8 +1,8 @@
-package EveryEgile.jaeyeon.repository;
+package org.everyagile.everyagile.everyagile.repository;
 
-import EveryEgile.jaeyeon.entity.Project;
+import org.everyagile.everyagile.everyagile.entity.Project;
 import org.springframework.data.jpa.repository.JpaRepository;
-import EveryEgile.jaeyeon.entity.Sprint;
+import org.everyagile.everyagile.everyagile.entity.Sprint;
 import org.springframework.data.repository.query.Param;
 
 import java.util.List;
