+++ src/main/resources/application.properties
@@ -1,1 +1,2 @@
-
+spring.h2.console.enabled=true
+spring.datasource.url=jdbc:h2:mem:testdb
