--- src/test/java/EveryEgile/jaeyeon/BoardApplicationTests.java
@@ -1,13 +1,0 @@
-package EveryEgile.jaeyeon;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class BoardApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
